Removed placeholder GameState and refactored controller to use empty Game class
